requirements.txt
Django==2.2
djangorestframework==3.8.2
django-debug-toolbar
django-jinja==2.4.1
django-querycount==0.7.0
django-cors-headers==2.4.0
django-filter==2.1.0
djangorestframework-jwt==1.11.0
django-rest-swagger==2.2.0
django-extensions
django-mptt==0.9.1
social-auth-app-django
Pillow==6.0.0
psycopg2-binary==2.8.1
python-social-auth
pytz==2018.9
boto3==1.7.10
django-background-tasks==1.2.0
firebase-admin==2.12.0
gunicorn==19.9.0
xlrd==1.2.0
openpyxl==2.6.2
sentry-sdk==0.10.2
django-ses
weasyprint
#pycairo
#pdfkit
# pdfkit depends on => sudo apt-get install wkhtmltopdf
#leaflet
#geojson
#Shapely
#python-pygraphviz
#pydotplus